LIME-124: * Separate enums
