+++ frontend/src/bundles/overview/components/circle-progress/enums/circle-progress-types.enums.ts
@@ -1,0 +1,8 @@
+const CircleProgressTypes = {
+    BASE: 'base',
+    RUNNING: 'running',
+    CYCLING: 'cycling',
+    WALKING: 'walking',
+};
+
+export { CircleProgressTypes };
