+++ frontend/src/bundles/overview/components/circle-progress/enums/enums.ts
@@ -1,0 +1,3 @@
+export { CircularColors as CircularProgressColors } from './circle-progress-color.enums.js';
+export { CircularSizes as CircularProgressSizes } from './circle-progress-sizes.enums.js';
+export { type CircleSizes } from './circle-progress-sizes.enums.js';
