--- frontend/src/bundles/overview/enums/enums.ts
@@ -1,5 +1,0 @@
-export {
-    CircularColors as CircularProgressColors,
-    CircularSizes as CircularProgressSizes,
-} from './circular-progress.enums.js';
-export { type CircleSizes } from './circular-progress.enums.js';
